+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 # Unreleased
 
+- Allow version extension via on_resolve_element
 - Resolve remotes during project element import
 - Allow arbitrary classes within on_resolve_element
 - Ignore errors during storage symlink creation